savestate: harden BSON parser against malformed/malicious states
Three related defects in the BSON deserialization layer:

1. bson_read_int32_array dereferences a NULL pointer when the requested
   key is missing. After the 'if (p) { ... }' early-return path, the
   function falls through to '*value = bson_read_u32(p);' with p == NULL,
   which is a guaranteed crash. The fall-through line is also pointless
   (the function is about to return false). Remove it.

2. bson_read_int32_array ignored its 'cnt' parameter and trusted the
   array element count encoded in the BSON document. A crafted savestate
   that claims a larger array than the caller allocated will overflow
   the destination buffer (e.g. reg[REG_ARCH_COUNT] in cpu_read_savestate,
   spsr[6], reg_mode[7*7], or rtc_data_array[2] in memory_read_savestate).
   Reject any state whose array size does not match the expected count.

3. bson_read_u32 expanded to '(p[3] << 24) | ...' where p is a u8*.
   Each element undergoes default integer promotion to signed int; for
   bytes >= 0x80, the '<< 24' shift moves the value into the sign bit
   of int, which is undefined behavior per the C standard. Cast to u32
   before shifting.

+++ savestate.h
@@ -36,8 +36,8 @@
 }
 
 #define bson_read_u32(p)                        \
-  ((p[3] << 24) | (p[2] << 16) |                \
-   (p[1] <<  8) | (p[0] <<  0))
+  (((u32)(p)[3] << 24) | ((u32)(p)[2] << 16) |  \
+   ((u32)(p)[1] <<  8) | ((u32)(p)[0] <<  0))
 
 #define bson_write_cstring(p, value)            \
   memcpy(p, value, strlen(value)+1);            \
